Move buffer.h to old_buffer.h and add a shim buffer.h
